remove ads; add about page
