+++ package.json
@@ -17,7 +17,6 @@
   },
   "private": true,
   "dependencies": {
-    "@admob-plus/ionic": "^1.1.1",
     "@angular/animations": "~12.2.1",
     "@angular/cdk": "^11.2.6",
     "@angular/common": "~12.2.1",
@@ -76,7 +75,6 @@
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "admob-plus-cordova": "^1.5.0",
     "codelyzer": "^6.0.0",
     "cordova-open-native-settings": "^1.5.3",
     "cordova-plugin-androidx-adapter": "^1.1.3",
@@ -127,11 +125,6 @@
       "mx.ferreyra.callnumber": {},
       "cordova-plugin-contacts": {},
       "cordova-sms-plugin": {},
-      "admob-plus-cordova": {
-        "APP_ID_ANDROID": "ca-app-pub-1258868559061405~8296554504",
-        "APP_ID_IOS": "test",
-        "PLAY_SERVICES_VERSION": "19.8.0"
-      },
       "cordova-plugin-theme-detection": {},
       "cordova-plugin-app-version": {},
       "cordova-open-native-settings": {},
