+++ src/app/pages/about/about.page.html
@@ -1,0 +1,93 @@
+<ion-header [translucent]="true" mode="ios">
+  <ion-toolbar color="dark" mode="ios">
+    <ion-buttons slot="start">
+      <ion-back-button mode="md" defaultHref="scan" [icon]="'arrow-back'">
+      </ion-back-button>
+    </ion-buttons>
+    <ion-title>{{ 'ABOUT' | translate }}</ion-title>
+  </ion-toolbar>
+</ion-header>
+
+<ion-content>
+
+  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" (click)="openRepoUrl()">
+    <ion-icon class="ion-margin-start ion-padding-horizontal"
+      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="logo-github">
+    </ion-icon>
+    <ion-label>
+      <p class="ion-padding">
+        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
+          {{ 'VIEW_SOURCE' | translate }}
+        </ion-text>
+        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
+          src="assets/icon/external-link.svg">
+        </ion-icon>
+      </p>
+    </ion-label>
+  </ion-item>
+  <ion-item *ngIf="isAndroid" mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" button
+    (click)="openGooglePlay()">
+    <ion-icon class="ion-margin-start ion-padding-horizontal"
+      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="logo-google-playstore">
+    </ion-icon>
+    <ion-label>
+      <p class="ion-padding">
+        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
+          {{ 'CHECK_GOOGLE_PLAY' | translate }}
+        </ion-text>
+        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
+          src="assets/icon/external-link.svg">
+        </ion-icon>
+      </p>
+    </ion-label>
+    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
+  </ion-item>
+  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" button
+    (click)="showUpdateNotes()">
+    <ion-icon class="ion-margin-start ion-padding-horizontal"
+      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" src="assets/icon/patch-update.svg">
+    </ion-icon>
+    <ion-label>
+      <p class="ion-padding">
+        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
+          {{ 'UPDATE_NOTES' | translate }}
+        </ion-text>
+      </p>
+    </ion-label>
+    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
+  </ion-item>
+  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" (click)="reportBug()">
+    <ion-icon class="ion-margin-start ion-padding-horizontal"
+      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="bug">
+    </ion-icon>
+    <ion-label>
+      <p class="ion-padding">
+        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
+          {{ 'REPORT_ISSUE' | translate }}
+        </ion-text>
+        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
+          src="assets/icon/external-link.svg">
+        </ion-icon>
+      </p>
+    </ion-label>
+    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
+  </ion-item>
+  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" button
+    (click)="viewPrivacyPolicy()">
+    <ion-icon class="ion-margin-start ion-padding-horizontal"
+      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="document-text-sharp">
+    </ion-icon>
+    <ion-label>
+      <p class="ion-padding">
+        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
+          {{ 'PRIVACY_POLICY' | translate }}
+        </ion-text>
+        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
+          src="assets/icon/external-link.svg">
+        </ion-icon>
+      </p>
+    </ion-label>
+    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
+  </ion-item>
+  
+</ion-content>
