+++ src/app/pages/setting/setting.page.html
@@ -75,98 +75,16 @@
     </ion-label>
     <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
   </ion-item>
-  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" (click)="openRepoUrl()">
+  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" (click)="navigateAbout()">
     <ion-icon class="ion-margin-start ion-padding-horizontal"
-      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="logo-github">
-    </ion-icon>
+      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="information-circle-sharp"></ion-icon>
     <ion-label>
       <p class="ion-padding">
         <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
-          {{ 'ABOUT_SIMPLE_QR' | translate }}
-        </ion-text>
-        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
-          src="assets/icon/external-link.svg">
-        </ion-icon>
-      </p>
-    </ion-label>
-    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
-  </ion-item>
-  <ion-item *ngIf="isAndroid" mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" button
-    (click)="openGooglePlay()">
-    <ion-icon class="ion-margin-start ion-padding-horizontal"
-      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="logo-google-playstore">
-    </ion-icon>
-    <ion-label>
-      <p class="ion-padding">
-        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
-          {{ 'CHECK_GOOGLE_PLAY' | translate }}
-        </ion-text>
-        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
-          src="assets/icon/external-link.svg">
-        </ion-icon>
-      </p>
-    </ion-label>
-    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
-  </ion-item>
-  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" button
-    (click)="showUpdateNotes()">
-    <ion-icon class="ion-margin-start ion-padding-horizontal"
-      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" src="assets/icon/patch-update.svg">
-    </ion-icon>
-    <ion-label>
-      <p class="ion-padding">
-        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
-          {{ 'UPDATE_NOTES' | translate }}
+          {{ 'ABOUT' | translate }}
         </ion-text>
       </p>
     </ion-label>
-    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
-  </ion-item>
-  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" button
-    (click)="viewPrivacyPolicy()">
-    <ion-icon class="ion-margin-start ion-padding-horizontal"
-      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="document-text-sharp">
-    </ion-icon>
-    <ion-label>
-      <p class="ion-padding">
-        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
-          {{ 'PRIVACY_POLICY' | translate }}
-        </ion-text>
-        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
-          src="assets/icon/external-link.svg">
-        </ion-icon>
-      </p>
-    </ion-label>
-    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
-  </ion-item>
-  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" (click)="reportBug()">
-    <ion-icon class="ion-margin-start ion-padding-horizontal"
-      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" name="bug">
-    </ion-icon>
-    <ion-label>
-      <p class="ion-padding">
-        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
-          {{ 'REPORT_ISSUE' | translate }}
-        </ion-text>
-        <ion-icon class="ion-padding-start" [color]="env.colorTheme === 'dark'? 'light' : 'dark'"
-          src="assets/icon/external-link.svg">
-        </ion-icon>
-      </p>
-    </ion-label>
-    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
-  </ion-item>
-  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" button
-    (click)="supportDeveloper()">
-    <ion-icon class="ion-margin-start ion-padding-horizontal"
-      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" src="assets/icon/dollar.svg"></ion-icon>
-    <ion-label>
-      <p class="ion-padding">
-        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
-          {{ 'SUPPORT_DEVELOPER' | translate }}
-        </ion-text>
-      </p>
-    </ion-label>
-    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
   </ion-item>
   <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="none" (click)="resetApp()">
     <ion-icon class="ion-margin-start ion-padding-horizontal"
