+++ src/app/pages/setting/setting.page.ts
@@ -49,36 +49,8 @@ export class SettingPage {
     this.router.navigate(['setting-search-engine']);
   }
 
-  openRepoUrl(): void {
-    window.open(this.env.GITHUB_REPO_URL, '_system');
-  }
-
-  openGooglePlay(): void {
-    window.open(this.env.GOOGLE_PLAY_URL, '_system');
-  }
-
-  async showUpdateNotes() {
-    const alert = await this.alertController.create({
-      header: this.translate.instant("UPDATE_NOTES"),
-      subHeader: this.env.appVersionNumber,
-      message: this.translate.instant("UPDATE.UPDATE_NOTES"),
-      buttons: [this.translate.instant("OK")],
-      cssClass: 'left-align'
-    });
-    await alert.present();
-  }
-
-  viewPrivacyPolicy(): void {
-    window.open(this.env.PRIVACY_POLICY, '_system');
-  }
-
-  async supportDeveloper() {
-    this.router.navigate(['support-developer']);
-  }
-
-  async reportBug() {
-    const mailContent = await this.env.getBugReportMailContent();
-    window.open(mailContent, '_system');
+  navigateAbout() {
+    this.router.navigate(['about']);
   }
 
   async resetApp() {
