--- src/app/pages/support-developer/support-developer.page.html
@@ -1,39 +1,0 @@
-<ion-header [translucent]="true" mode="ios">
-  <ion-toolbar color="dark" mode="ios">
-    <ion-buttons slot="start">
-      <ion-back-button mode="md" defaultHref="scan" [icon]="'arrow-back'">
-      </ion-back-button>
-    </ion-buttons>
-    <ion-title> {{ 'SUPPORT_DEVELOPER' | translate }}</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-
-  <ion-item mode="ios" class="ion-no-padding ripple-parent" detail="false" lines="full" (click)="watchAds()">
-    <ion-icon class="ion-margin-start ion-padding-horizontal"
-      [color]="env.colorTheme === 'dark'? 'primary' : 'secondary'" src="assets/icon/ads.svg"></ion-icon>
-    <ion-label>
-      <p class="ion-padding">
-        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="font-size: large;">
-          {{ 'WATCH_ADS' | translate }}
-        </ion-text>
-      </p>
-    </ion-label>
-    <!--<ion-ripple-effect type="bounded"></ion-ripple-effect> -->
-  </ion-item>
-  <ion-item class="ion-no-padding" lines="none">
-    <ion-label class="ion-padding-start">
-      <p class="ion-padding">
-        <ion-text [color]="env.colorTheme === 'dark'? 'light' : 'dark'" style="white-space: normal;"
-          [innerHTML]="'SUPPORT_DEV_EXPLAIN' | translate">
-        </ion-text>
-      </p>
-    </ion-label>
-  </ion-item>
-
-  <ion-row class="ion-padding-bottom">
-    <ion-col></ion-col>
-  </ion-row>
-
-</ion-content>
