+++ src/assets/i18n/en.json
@@ -7,7 +7,7 @@
     "CREATE_QRCODE": "Create QR Code",
     "CREATE_QRCODE_MAX_LENGTH": "Max. 1817 characters",
     "QRCODE_CONTENT": "QR code content",
-    "EXIT_APP": "Exit the app",
+    "EXIT_APP": "Exit App",
     "WEBSITE": "Website",
     "PHONE_CALL": "Phone Call",
     "SEARCH": "Search",
@@ -16,6 +16,7 @@
     "YES": "Yes",
     "NO": "No",
     "OK": "OK",
+    "ABOUT": "About",
     "ENTER": "Enter",
     "CONFIRM": "Confirm",
     "CANCEL": "Cancel",
@@ -92,7 +93,7 @@
     "SEE_MORE": "See More",
     "SUCCESS": "Success",
     "RESTORED": "Restored",
-    "SUPPORT_DEV_EXPLAIN": "Simple QR is a free and open-source app. If you would like to support me and my other projects, please feel free to watch ads. Thank you for your support!",
+    "SUPPORT_DEV_EXPLAIN": "Simple QR is a free and open-source app. If you want to support me and my projects, please feel free to contribute to the projects!",
     "DEVELOPING": "Developing",
     "REPORT_ISSUE": "Report Issue",
     "RESET_APP": "Reset App",
@@ -163,6 +164,7 @@
     "SCAN": "Scan",
     "IMPORT": "Import",
     "IMPORT_IMAGE": "Import Image",
+    "VIEW_SOURCE": "View Source",
     "MSG": {
         "CALL_PHONE": "Are you sure to call {phoneNumber} now?",
         "FAIL_CALL_PHONE": "Failed to open dialer",
