+++ src/assets/i18n/zh-HK.json
@@ -7,7 +7,7 @@
     "CREATE_QRCODE": "建立 QR 碼",
     "CREATE_QRCODE_MAX_LENGTH": "最多 1817 個字元",
     "QRCODE_CONTENT": "QR 碼內容",
-    "EXIT_APP": "離開",
+    "EXIT_APP": "離開應用程式",
     "WEBSITE": "網站",
     "PHONE_CALL": "通話",
     "SEARCH": "搜尋",
@@ -16,6 +16,7 @@
     "YES": "是",
     "NO": "否",
     "OK": "好的",
+    "ABOUT": "關於",
     "ENTER": "輸入",
     "CONFIRM": "確定",
     "CANCEL": "取消",
@@ -92,7 +93,7 @@
     "SEE_MORE": "查看更多",
     "SUCCESS": "成功",
     "RESTORED": "已恢復",
-    "SUPPORT_DEV_EXPLAIN": "簡易 QR 是一個免費的開源應用程式。如果您想支持開發人員，歡迎您觀看廣告。感謝您的支持！",
+    "SUPPORT_DEV_EXPLAIN": "簡易 QR 是一個免費的開源應用程式，歡迎您查閱源代碼及作出貢獻。",
     "DEVELOPING": "開發中",
     "REPORT_ISSUE": "回報問題",
     "RESET_APP": "重設應用程式",
@@ -163,6 +164,7 @@
     "SCAN": "掃描",
     "IMPORT": "匯入",
     "IMPORT_IMAGE": "匯入圖片",
+    "VIEW_SOURCE": "查閱源代碼",
     "MSG": {
         "CALL_PHONE": "確定要致電 {phoneNumber} ?",
         "FAIL_CALL_PHONE": "無法開啟電話",
